Delete generateWallet thunk
